+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchantments/enchantment/AirBorn.java
@@ -7,7 +7,6 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.enchantment.DiggingEnchantment;
 import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.item.enchantment.EnchantmentCategory;
 import net.minecraft.world.level.block.state.BlockState;
@@ -25,10 +24,6 @@ public class AirBorn extends Enchantment implements IEnchantmentTooltip {
     @Override
     public int getMaxCost(int level) {
         return this.getMinCost(level) + 50;
-    }
-
-    public boolean checkCompatibility(Enchantment other) {
-        return !(other instanceof DiggingEnchantment) && super.checkCompatibility(other);
     }
 
     public static float getMiningSpeedMultiplier(LivingEntity entity, BlockState state) {